+++ HISTORY.md
@@ -3,8 +3,13 @@
 https://github.com/josdejong/jsoneditor
 
 
-## not yet released, version 5.15.1
+## not yet released, version 5.16.0
 
+- Better handling of large JSON documents:
+  - Only displays the first 100 items of large arrays, with buttons
+    "show more" and "show all" to render more items.
+  - Search does no longer expand the paths to all matches, instead
+    it only expands the path of the current search result.
 - Fixed index numbers of Array items not being updated after sorting.